Add maximum message size
Sending invalid length data leads to problems:
- Out of memory for very large values
- Panic/crash for values less than 0

This fixes this by checking the size before trying to allocate the
buffer.

I chose 512 bytes as the maximum message size. If I saw it correctly,
the biggest request is CreateRoom, with a payload of about 120 bytes.
With a bit to spare, 512 seems to be a sensible choice.